+++ packages/vquery/rslib.config.ts
@@ -4,12 +4,29 @@ export default defineConfig({
   lib: [
     {
       format: 'esm',
-      syntax: ['node 18'],
+      source: {
+        entry: {
+          browser: './src/browser.ts',
+        },
+      },
+      output: {
+        distPath: './dist/esm',
+      },
       dts: true,
+      bundle: true,
     },
     {
       format: 'cjs',
-      syntax: ['node 18'],
+      source: {
+        entry: {
+          node: './src/node.ts',
+        },
+      },
+      output: {
+        distPath: './dist/cjs',
+      },
+      dts: true,
+      bundle: true,
     },
   ],
 })