+++ packages/vquery/src/sql-builder/utils.ts
@@ -1,5 +1,5 @@
 import { Where, WhereClause, WhereGroup, WhereLeaf } from 'src/types'
-import { SelectItem } from 'src/types/QueryDSL/Select'
+import { SelectItem } from 'src/types/dsl/Select'
 
 export const isSelectItem = <T>(item: keyof T | SelectItem<T>): item is SelectItem<T> => {
   return typeof item === 'object' && 'field' in item