+++ packages/vseed/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @visactor/vseed
 
+## 0.2.5
+
+## 0.2.4
+
 ## 0.2.3
 
 ### Patch Changes